Ajustes nas cores e icone principal do app

+++ app/src/main/java/com/produze/sistemas/vendasnow/vendasnowpremium/model/SaleProduct.kt
@@ -1,8 +1,8 @@
 package com.produze.sistemas.vendasnow.vendasnowpremium.model
 
 data class SaleProduct (
-        var quantity: Int = 0,
+        var quantity: Double = 0.00,
         var valueSale: Double = 0.00,
         var product: Product?) {
-    constructor():this(0,0.00, null)
+    constructor():this(0.00,0.00, null)
 }
