+++ app/src/main/java/com/produze/sistemas/vendasnow/vendasnowpremium/model/SaleService.kt
@@ -1,8 +1,8 @@
 package com.produze.sistemas.vendasnow.vendasnowpremium.model
 
 class SaleService (
-        var quantity: Int = 0,
+        var quantity: Double = 0.00,
         var valueSale: Double = 0.00,
         var service: Service?) {
-    constructor():this(0,0.00, null)
+    constructor():this(0.00,0.00, null)
 }
